homework 해야됨

+++ chap03/src/homework/Homework4.java
@@ -17,7 +17,8 @@ public class Homework4 {
 		int english = sc.nextInt();
 		
 		int total = korean+math+english;
-		int average = total / 3;
+		double average = total / 3.0; // 소숫점 결과를 얻고 싶다면 double로 바꿔주고
+		//나누는 것중 하나를 double형인 3.0으로 바꿈
 		System.out.println("합계: " + total + "점");
 		System.out.println("평균: " + average + "점");
 		
@@ -28,8 +29,10 @@ public class Homework4 {
 		
 		System.out.println(parentsSay);
 		// 250717_23:18에 쌤 깃허브 보고있는데 쌤은 
+		//boolean canChangePhone = (korean >= 60 && math >= 60 && english >= 60 && avg > 90) ?
+	    		//true : false;
 //		System.out.println(canChangePhone ? "휴대폰을 바꿀 수 있습니다." : "휴대폰을 바꿀 수 없습니다.");
-		// 이런식으로 syso 에다 조건을 삼항연산자를 걸어버림
+		// 이런식으로 boolean으로 true,false얻어내고 syso 에다 삼항연산자를 걸어버림
 	}
 
 }
